feat: 小图片处理
