+++ packages/chat-sdk/src/libs/ui-kit/atomic/md-stream/render/phrase/image/index.module.less
@@ -11,9 +11,9 @@
   text-decoration: inherit;
 
   &.image-error {
-  background-image: url(../../../../../assets/svg/image.svg);
-  background-position: center;
-  background-size: auto;
-  background-repeat: no-repeat;
-}
+    background-image: url(../../../../../assets/svg/image.svg);
+    background-position: center;
+    background-size: auto;
+    background-repeat: no-repeat;
+  }
 }
