+++ packages/chat-sdk/src/pages/chat/index.config.ts
@@ -1,0 +1,3 @@
+export default definePageConfig({
+  navigationBarTitleText: 'Coze Chat',
+});
