+++ packages/chat-sdk/src/pages/index/index.module.less
@@ -1,56 +1,12 @@
 .container {
-  position: absolute;
-  height: 100%;
   display: flex;
-  flex-direction: column;
-  background-color: #FFF;
-  overflow: hidden;
-  width: -webkit-fill-available;
-  box-sizing: border-box;
-  color: #000;
-}
+  gap: 10px;
+  padding: 20px;
 
-.chat-container {
-  flex: 1;
-  overflow: hidden;
-
-  .chat-slot {
-    padding: 0px 24px 21px;
-  }
-}
-
-.wave {
-  background: url('../../libs/ui-kit//assets//svg/audio/1-blue-fill.svg');
-  background-repeat: no-repeat;
-  background-size: cover;
-  width: 20px;
-  height: 20px;
-  animation: icons-blue 1.5s infinite;
-
-}
-
-@keyframes icons-blue {
-  0% {
-    background: url('../../libs/ui-kit//assets//svg/audio/1-blue-fill.svg');
-    background-repeat: no-repeat;
-    background-size: cover;
-  }
-
-  33% {
-    background: url('../../libs/ui-kit//assets//svg/audio/2-blue-fill.svg');
-    background-repeat: no-repeat;
-    background-size: cover;
-  }
-
-  66% {
-    background: url('../../libs/ui-kit//assets//svg/audio/3-blue-fill.svg');
-    background-repeat: no-repeat;
-    background-size: cover;
-  }
-
-  100% {
-    background: url('../../libs/ui-kit//assets//svg/audio/1-blue-fill.svg');
-    background-repeat: no-repeat;
-    background-size: cover;
+  .nav {
+    cursor: pointer;
+    background: #aaa;
+    border-radius: 6px;
+    padding: 0 20px;
   }
 }
