+++ packages/chat-sdk/src/pages/markdown/const.ts
@@ -9,6 +9,13 @@ $(document).ready(function () {
 
 [Google](https://www.google.com)
 ![Alt Text](https://pic1.zhimg.com/v2-b444070848d54baf536222b22a51fba4_b.jpg)
+
+![Alt Text](https://lf3-static.bytednsdoc.com/obj/eden-cn/vhnupclkeh7nupkvhn/20250221-175449.jpeg)
+
+
+![Alt Text](https://lf3-static.bytednsdoc.com/obj/eden-cn/rkzild_lgvj/ljhwZthlaukjlkulzlp/assets/imgs/coze-logo.png)
+![Alt Text](https://lf3-static.bytednsdoc.com/obj/eden-cn/vhnupclkeh7nupkvhn/3bdef16c-3d10-43d0-9c4f-123abbee3a42.png)
+
 as s \`node\`
 ## Html
