+++ packages/chat-sdk/src/pages/markdown/index.tsx
@@ -18,17 +18,21 @@ const streamOutput = (onChange: (content: string) => void) => {
   }, 150);
 };
 
+const isSmooth = false;
 export default function Index() {
-  const [content, setContent] = useState('');
+  const [content, setContent] = useState(isSmooth ? markdown : '');
 
   useEffect(() => {
+    if (isSmooth) {
+      return;
+    }
     streamOutput(setContent);
   }, []);
 
   return (
     <View className="light">
       <MdStream
-        isSmooth
+        isSmooth={isSmooth}
         isFinish={content.length === markdown.length}
         enableCodeBy4Space={false}
         markdown={content}
